requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
openai==1.12.0
pydub==0.25.1
ffmpeg-python==0.2.0 